refactor(cast): inline disassemble implementation
